+++ examples/common/common-pmem_map_file.h
@@ -18,6 +18,9 @@ struct common_mem {
 	char *mr_ptr;
 	/* size of the mapped persistent memory */
 	size_t mr_size;
+	/* offset of the beginning of the used persistent memory */
+	size_t offset;
+	/* offset of user data after the pmem signature */
 	size_t data_offset;
 	int is_pmem;
 	persist_fn persist;